Move builder related classes into its own package
