+++ src/main/java/com/android/volley/toolbox/StringRequest.java
@@ -21,6 +21,7 @@ import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.Response.ErrorListener;
 import com.android.volley.Response.Listener;
+import com.android.volley.toolbox.requestbuilder.ResponseParsers;
 
 /** A canned request for retrieving the response body at a given URL as a String. */
 public class StringRequest extends Request<String> {
