+++ src/main/java/com/android/volley/toolbox/requestbuilder/BuildableRequest.java
@@ -1,6 +1,6 @@
-package com.android.volley.toolbox;
+package com.android.volley.toolbox.requestbuilder;
 
-import static com.android.volley.toolbox.Utils.requireNonNull;
+import static com.android.volley.Utils.requireNonNull;
 
 import android.net.Uri;
 import com.android.volley.AuthFailureError;
