+++ src/main/java/com/android/volley/toolbox/requestbuilder/ResponseParser.java
@@ -1,4 +1,4 @@
-package com.android.volley.toolbox;
+package com.android.volley.toolbox.requestbuilder;
 
 import com.android.volley.NetworkResponse;
 import com.android.volley.Response;
